Enhance Solaris Packaging -- #2613
This is the beginning of splitting out the IDO facility from the
base Icinga component to facilitate packaging for Solaris.  This
will allow folks who cannot get over the "dependency hump" to
install an IDO-free instance and at least get the benefit of the
classic CGIs and monitoring.

The new Makefile.in is smart enough to suppress the installation
of the IDO bits into Solaris package if IDO is not in use and to
provide it if it is.

I'm committing at the moment so I don't forget and check something
else out and blow away my changes.

refs #2613